fmt: drop cast before new->len (clang-format 18/23 disagree on spacing)
clang-format 18 fails to parse '(size_t)new->len' as a C cast (the
'new' token) and forces a space; 23.1.0 (CI) flags that space as a
violation — the two versions fight over this exact construct. int->
size_t widening is silent under -Wall -Wextra, so drop the cast and
remove the construct entirely.

+++ src/str.c
@@ -426,7 +426,7 @@ static Val str_replace(VM *vm, Val *args, int nargs) {
     int off = 0, i = 0;
     while (i < s->len) {
         if (i <= s->len - old->len && memcmp(s->data + i, old->data, (size_t)old->len) == 0) {
-            memcpy(tmp + off, new->data, (size_t) new->len);
+            memcpy(tmp + off, new->data, new->len);
             off += new->len;
             i += old->len;
         } else {
